Share view logic in PaywallCoordinator
